LPS-56182 Move classes
