+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-io/src/com/liferay/dynamic/data/mapping/io/internal/DDMFormLayoutJSONSerializerImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.io.impl;
+package com.liferay.dynamic.data.mapping.io.internal;
 
 import com.liferay.dynamic.data.mapping.io.DDMFormLayoutJSONSerializer;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
